refactor(notes): move InsecureContextBanner to src/components/
Closes #143. The banner was first defined in src/app/routes/AddVault.tsx
where it was originally used, then imported from there by VaultPopover
and VaultStatusBanner — a components-importing-from-routes dependency
inversion. Relocate to src/components/InsecureContextBanner.tsx with no
behavior change; update three import sites.

- Move definition to src/components/InsecureContextBanner.tsx
- AddVault.tsx now imports it from @/components/InsecureContextBanner
- VaultPopover.tsx and VaultStatusBanner.tsx switch their import path
  to @/components/InsecureContextBanner

Bump 0.3.16-rc.1 → 0.3.16-rc.2 + CHANGELOG.

813/813 tests pass; typecheck, lint, build clean.
